# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SafeSpace Backend",
    "description": "Secure anonymous chatbot backend with PostgreSQL persistence",
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/sessions": {
      "post": {
        "tags": ["Sessions"],
        "summary": "Create a new anonymous session",
        "description": "Creates a new anonymous chat session with a unique UUID. Returns session ID and creation timestamp.",
        "operationId": "create_session_api_v1_sessions_post",
        "responses": {
          "201": {
            "description": "Session created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SessionResponse" }
              }
            }
          },
          "503": { "description": "Database unavailable" }
        }
      }
    },
    "/api/v1/sessions/{session_id}": {
      "delete": {
        "tags": ["Sessions"],
        "summary": "Delete a session",
        "description": "Permanently deletes a session and all associated data. This action cannot be undone.",
        "operationId": "delete_session_api_v1_sessions__session_id__delete",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "204": { "description": "Session deleted successfully" },
          "503": { "description": "Database unavailable" },
          "404": { "description": "Session not found" },
          "422": { "description": "Invalid session ID format" }
        }
      }
    },
    "/api/v1/sessions/{session_id}/chat": {
      "post": {
        "tags": ["Chat"],
        "summary": "Send Message",
        "operationId": "send_message_api_v1_sessions__session_id__chat_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MessageCreate" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MessageResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Chat"],
        "summary": "Get Chat History",
        "operationId": "get_chat_history_api_v1_sessions__session_id__chat_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ChatHistoryResponse" }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HTTPValidationError" }
              }
            }
          }
        }
      }
    },
    "/": {
      "get": {
        "summary": "Read Root",
        "description": "Root endpoint - health indicator.",
        "operationId": "read_root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": { "type": "string" },
                  "type": "object",
                  "title": "Response Read Root  Get"
                }
              }
            }
          }
        }
      }
    },
    "/health/db": {
      "get": {
        "summary": "Health Check Db",
        "description": "Database health check endpoint.\nVerifies database connectivity and returns connection pool status.\n\nReturns:\n    - status: \"healthy\" if database is accessible, \"unhealthy\" otherwise\n    - database: \"connected\" or \"disconnected\"\n    - timestamp: ISO format timestamp\n    - connection_pool: Current pool statistics\n    - error: (optional) Error message if unhealthy",
        "operationId": "health_check_db_health_db_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Health Check Db Health Db Get"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ChatHistoryResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "messages": {
            "items": { "$ref": "#/components/schemas/MessageResponse" },
            "type": "array",
            "title": "Messages"
          }
        },
        "type": "object",
        "required": ["session_id", "messages"],
        "title": "ChatHistoryResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": { "$ref": "#/components/schemas/ValidationError" },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MessageCreate": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 4096,
            "minLength": 1,
            "title": "Message",
            "description": "Pesan dari user"
          }
        },
        "type": "object",
        "required": ["message"],
        "title": "MessageCreate"
      },
      "MessageResponse": {
        "properties": {
          "id": { "type": "string", "format": "uuid", "title": "Id" },
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "description": "'user' atau 'model'"
          },
          "content": { "type": "string", "title": "Content" },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": ["id", "session_id", "role", "content", "created_at"],
        "title": "MessageResponse"
      },
      "SessionResponse": {
        "properties": {
          "session_id": {
            "type": "string",
            "format": "uuid",
            "title": "Session Id",
            "description": "Unique session identifier (UUID4)",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Session creation timestamp in ISO 8601 format",
            "example": "2026-02-21T13:30:45.123456"
          }
        },
        "type": "object",
        "required": ["session_id", "created_at"],
        "title": "SessionResponse",
        "description": "Anonymous user session response",
        "example": {
          "created_at": "2026-02-21T13:30:45.123456",
          "session_id": "550e8400-e29b-41d4-a716-446655440000"
        }
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": { "anyOf": [{ "type": "string" }, { "type": "integer" }] },
            "type": "array",
            "title": "Location"
          },
          "msg": { "type": "string", "title": "Message" },
          "type": { "type": "string", "title": "Error Type" }
        },
        "type": "object",
        "required": ["loc", "msg", "type"],
        "title": "ValidationError"
      }
    }
  }
}